refactor(server): rename trailingDebounce.ts to TrailingDebouncer.ts
Files whose primary export is a class follow PascalCase in this
codebase (BaseParser.ts, GCodeScanner.ts, WorkspaceIndexingService.ts,
etc.). `trailingDebounce.ts` exported the `TrailingDebouncer` class
and was the only camelCase outlier introduced in #138. Renames the
production file and its test to match the convention and updates
the two import sites.

Pure rename + import update — no behavior change.

Refs #138
